requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
aiosqlite
pydantic
python-dotenv
reportlab
pydantic[email]
jinja2
aiosmtplib
bcrypt
PyJWT